refactor: extract config and constants to new module
