+++ config/go.mod
@@ -1,0 +1,5 @@
+module github.com/hanapedia/the-bench/config
+
+go 1.20
+
+require gopkg.in/yaml.v3 v3.0.1
