+++ config/yaml/config.go
@@ -1,9 +1,9 @@
-package config
+package yaml
 
 import (
 	"os"
 
-	"github.com/hanapedia/the-bench/service-unit/internal/domain/core"
+	"github.com/hanapedia/the-bench/config/model"
 	"gopkg.in/yaml.v3"
 )
 
@@ -11,10 +11,10 @@ type YamlConfigLoader struct {
 	Path string
 }
 
-func (ycl YamlConfigLoader) Load() (core.ServiceUnitConfig, error) {
+func (ycl YamlConfigLoader) Load() (model.ServiceUnitConfig, error) {
 	data, err := os.ReadFile(ycl.Path)
 
-	var config core.ServiceUnitConfig
+	var config model.ServiceUnitConfig
 	err = yaml.Unmarshal(data, &config)
 
     return config, err
