+++ service-unit/go.mod
@@ -4,8 +4,9 @@ go 1.20
 
 require (
 	github.com/gofiber/fiber/v2 v2.43.0
+	github.com/hanapedia/the-bench/config v0.0.0-00010101000000-000000000000
 	github.com/segmentio/kafka-go v0.4.39
-	gopkg.in/yaml.v3 v3.0.1
+	go.mongodb.org/mongo-driver v1.11.4
 )
 
 require (
@@ -31,9 +32,11 @@ require (
 	github.com/xdg-go/scram v1.1.1 // indirect
 	github.com/xdg-go/stringprep v1.0.3 // indirect
 	github.com/youmark/pkcs8 v0.0.0-20181117223130-1be2e3e5546d // indirect
-	go.mongodb.org/mongo-driver v1.11.4 // indirect
 	golang.org/x/crypto v0.7.0 // indirect
 	golang.org/x/sync v0.1.0 // indirect
 	golang.org/x/sys v0.6.0 // indirect
 	golang.org/x/text v0.8.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
+
+replace github.com/hanapedia/the-bench/config => ../config
