--- service-unit/internal/domain/core/service_unit.go
@@ -1,1 +1,0 @@
-package core
