+++ service-unit/internal/infrastructure/egress/factory/mongoAdapterFactory.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/hanapedia/the-bench/service-unit/internal/domain/core"
 	"github.com/hanapedia/the-bench/service-unit/internal/infrastructure/egress/repository_adapter/mongo"
-	"github.com/hanapedia/the-bench/service-unit/pkg/constants"
+	"github.com/hanapedia/the-bench/config/constants"
 )
 
 func (egressAdapterDetails EgressAdapterDetails) mongoEgressAdapterFactory() (core.EgressAdapter, error) {
