+++ service-unit/internal/infrastructure/egress/invocation_adapter/rest/write.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/hanapedia/the-bench/service-unit/internal/domain/contract"
-	"github.com/hanapedia/the-bench/service-unit/pkg/constants"
+	"github.com/hanapedia/the-bench/config/constants"
 	"github.com/hanapedia/the-bench/service-unit/pkg/utils"
 )
